+++ src/Checkbox.js
@@ -1,5 +1,4 @@
 import React from "react";
-// import "./scss/main.scss";
 
 function CheckBox(props) {
   return (
